support for hybrid transactions
Hybrid transactions are those where some signatures come from a
hardware wallet and some come from a cold wallet. The app will
witness the transaction using the hardware wallet keys, and then
export both the new witness files and the tx.body file. The tx
can then be witnessed with the cold key, and all of the witnesses
can be assembled to produce the final tx.signed file. This
tx.signed file can be imported into the app for submission to the
blockchain.