dynamically calculate the required collateral
During the build step, the `collateralPercentage` network
parameter is used to calculate the required collateral amount
based on the calculated transaction fee.

All collateral inputs are required to contain at least 5 ADA. This
may be an issue at some point in the future, but for now, this
requirement is hard-coded into the app. The change will be returned
to the address the collateral input belongs to.