added support for outputs to external addresses + builder tweak
Previously, the only way to create an output was to pay someone
in your address book. Now, an external output can be added directly
from the builder.

The builder notification dot in the main menu now turns blue when
the transaction can be successfully built. Previously, it was always
red.