fixed weird scaling for certain monitor resolutions
The monomer auto-scaling would create a small window with small text
for some screen resolutions. There doesn't seem to be an easy fix
for the auto-scaling so instead, the scaling is fixed to 2 and the
window is always open maximized. This isn't ideal since it doesn't
adapt to the resolution of the monitor, but it is good enough for
the prototype.