fix: pool picker stuck in a loop (closes #4)
While exploring Issue #4, it was discovered that Koios was consistently
returning a gateway timeout error when using its `pool_info`
[endpoint](https://api.koios.rest/#post-/pool_info). This endpoint was
being used to get the "live" stake pool stats such as `live_saturation`
and `live_pledge`. However, after
[talking](cardano-community/koios-artifacts#310) to the Koios team, it
was decided not to use the `pool_info` endpoint for the pool picker
because it is too heavy of a query to do in bulk (ie, for all registered
pools). Instead, the lighter `pool_list`
[endpoint](https://api.koios.rest/#get-/pool_list) was expanded to
include `active_stake` so that it could be used instead.

The net impacts for this are:
1. The pool picker no longer gets stuck in a loop.
2. Loading the pool picker is much faster.
3. Live saturation and live pledge are no longer available in the pool picker.
4. The pool picker is more tightly designed around Koios.

Number 3 can be addressed by adding a "see more details" button for the
pool in question. (This feature will be addressed in a separate
commit/PR).

Number 4 does not have a simple solution. However, looking at what
[blockfrost](https://docs.blockfrost.io/#tag/cardano--pools/GET/pools/extended)
and
[maestro](https://docs.gomaestro.org/cardano/list-registered-stake-pools)
offer, they don't even seem to have an equivalent to Koios' `pool_list`
endpoint which means they likely can't be used for the pool picker
anyway.