fix: sale and bid information wasn't scrollable
Not being able to scroll prevented closing certain menus. Unfortunately,
making them scrollable ruined the centering of the menus. AFAIK there is
no way around this as long as the current version of the Haskell Monomer
library is being used. So it now works on low resolution screens, but
isn't centered on high resolution screens.