feat: added ability to specify custom deposits for Claim Bids
Since the bid deposit is used to compensate the seller in the case where
the bidder walks away after the seller accepts the bid, a larger bid
deposit would make sellers more likely to accept this risk and take
Claim Bids seriously. If the bidder is committed, a large deposit can
convey this commitment. The bidder will get the deposit back as long as
they actually claim the NFTs.